v0.18 Single-instance environment encapsulate successfully
July 3rd:

Today, I finally set up a single-instance environment successfully.  In
the end of each trial, the environment is able to restart the same trial
nicely and smoothly. When it reaches the excution number, the
environment jumps back to menu.

The agent hasn't been programmed yet. The reward system is crude (return
50 points when it lands on a hill). Way to go!